chore: move init provider function to provider module
