+++ wren-ai-service/src/pipelines/retrieval/retrieval.py
@@ -356,7 +356,7 @@ if __name__ == "__main__":
 
     from src.core.engine import EngineConfig
     from src.core.pipeline import async_validate
-    from src.utils import init_langfuse, init_providers, load_env_vars
+    from src.providers import init_langfuse, init_providers, load_env_vars
 
     load_env_vars()
     init_langfuse()
