+++ wren-ai-service/tests/pytest/pipelines/test_ask_details.py
@@ -2,7 +2,7 @@ import pytest
 
 from src.core.engine import EngineConfig
 from src.pipelines.generation import sql_breakdown
-from src.utils import init_providers
+from src.providers import init_providers
 
 
 @pytest.mark.asyncio
